refactor(embed-examples): use scoped named data updates
