+++ packages/embed-examples/src/dynamicNamedData.html
@@ -8,7 +8,10 @@
 
     <body>
         <h1>Dynamic Named Data</h1>
-        <p>New data is pushed using the <code>updateNamedData</code> method.</p>
+        <p>
+            New data is pushed using the owning view's
+            <code>setNamedData</code> method.
+        </p>
         <div id="container"></div>
         <script type="module" src="/dynamicNamedData.js"></script>
     </body>
