+++ packages/embed-examples/src/index.html
@@ -40,7 +40,7 @@
                     >Multiple dynamic sources with a shared scale domain</a
                 >
             </li>
-            <li><a href="namedDataProvider.html">Named data provider</a></li>
+            <li><a href="namedDataForm.html">Named data form</a></li>
             <li><a href="dynamicFasta.html">Dynamic fasta experiment</a></li>
             <li><a href="reactComponent.html">React component example</a></li>
         </ul>
